refactor: dashboard count 수정
회원의 활동 카운트는 각각 게시물 목록을 불러와서 페이지 요청마다 5개의 쿼리를 요청하는데 수정한 쿼리문은 한번만 요청한다

+++ src/main/java/com/phcworld/domain/user/DashBoardUser.java
@@ -13,13 +13,13 @@ public class DashBoardUser {
 	
 	private User user;
 	
-	private Integer countOfAnswer;
+	private Long countOfAnswer;
 	
-	private Integer countOfFreeBoard;
+	private Long countOfFreeBoard;
 	
-	private Integer countOfDiary;
+	private Long countOfDiary;
 	
-	private Integer countOfAlert;
+	private Long countOfAlert;
 	
 	private List<Timeline> timelineList;
 }
